Moving everything into odc.geo.
